+++ Firestore/Example/Tests/Integration/API/FIRQueryTests.mm
@@ -21,6 +21,7 @@
 #import "Firestore/Source/API/FIRQuery+Internal.h"
 
 #import "Firestore/Example/Tests/Util/FSTEventAccumulator.h"
+#import "Firestore/Example/Tests/Util/FSTHelpers.h"
 #import "Firestore/Example/Tests/Util/FSTIntegrationTestCase.h"
 
 @interface FIRQueryTests : FSTIntegrationTestCase
@@ -54,6 +55,84 @@
 
   XCTAssertEqualObjects(FIRQuerySnapshotGetData(snapshot),
                         (@[ @{@"k" : @"d", @"sort" : @2}, @{@"k" : @"c", @"sort" : @1} ]));
+}
+
+- (void)testLimitToLastMustAlsoHaveExplicitOrderBy {
+  FIRCollectionReference *collRef = [self collectionRefWithDocuments:@{}];
+  FIRQuery *query = [collRef queryLimitedToLast:2];
+  FSTAssertThrows(
+      [query getDocumentsWithCompletion:^(FIRQuerySnapshot *documentSet, NSError *error){
+      }],
+      @"limit(toLast:) queries require specifying at least one OrderBy() clause.");
+}
+
+// Two queries that mapped to the same target ID are referred to as
+// "mirror queries". An example for a mirror query is a limitToLast()
+// query and a limit() query that share the same backend Target ID.
+// Since limitToLast() queries are sent to the backend with a modified
+// orderBy() clause, they can map to the same target representation as
+// limit() query, even if both queries appear separate to the user.
+- (void)testListenUnlistenRelistenSequenceOfMirrorQueries {
+  FIRCollectionReference *collRef = [self collectionRefWithDocuments:@{
+    @"a" : @{@"k" : @"a", @"sort" : @0},
+    @"b" : @{@"k" : @"b", @"sort" : @1},
+    @"c" : @{@"k" : @"c", @"sort" : @1},
+    @"d" : @{@"k" : @"d", @"sort" : @2},
+  }];
+
+  // Setup a `limit` query.
+  FIRQuery *limit = [[collRef queryOrderedByField:@"sort" descending:NO] queryLimitedTo:2];
+  FSTEventAccumulator *limitAccumulator = [FSTEventAccumulator accumulatorForTest:self];
+  id<FIRListenerRegistration> limitRegistration =
+      [limit addSnapshotListener:limitAccumulator.valueEventHandler];
+
+  // Setup a mirroring `limitToLast` query.
+  FIRQuery *limitToLast = [[collRef queryOrderedByField:@"sort"
+                                             descending:YES] queryLimitedToLast:2];
+  FSTEventAccumulator *limitToLastAccumulator = [FSTEventAccumulator accumulatorForTest:self];
+  id<FIRListenerRegistration> limitToLastRegistration =
+      [limitToLast addSnapshotListener:limitToLastAccumulator.valueEventHandler];
+
+  // Verify both queries get expected result.
+  FIRQuerySnapshot *snapshot = [limitAccumulator awaitEventWithName:@"Snapshot"];
+  NSArray *expected = @[ @{@"k" : @"a", @"sort" : @0}, @{@"k" : @"b", @"sort" : @1} ];
+  XCTAssertEqualObjects(FIRQuerySnapshotGetData(snapshot), expected);
+  snapshot = [limitToLastAccumulator awaitEventWithName:@"Snapshot"];
+  expected = @[ @{@"k" : @"b", @"sort" : @1}, @{@"k" : @"a", @"sort" : @0} ];
+  XCTAssertEqualObjects(FIRQuerySnapshotGetData(snapshot), expected);
+
+  // Unlisten then re-listen limit query.
+  [limitRegistration remove];
+  [limit addSnapshotListener:[limitAccumulator valueEventHandler]];
+
+  // Verify limit query still works.
+  snapshot = [limitAccumulator awaitEventWithName:@"Snapshot"];
+  expected = @[ @{@"k" : @"a", @"sort" : @0}, @{@"k" : @"b", @"sort" : @1} ];
+  XCTAssertEqualObjects(FIRQuerySnapshotGetData(snapshot), expected);
+
+  // Add a document that would change the result set.
+  [self addDocumentRef:collRef data:@{@"k" : @"e", @"sort" : @-1}];
+
+  // Verify both queries get expected result.
+  snapshot = [limitAccumulator awaitEventWithName:@"Snapshot"];
+  expected = @[ @{@"k" : @"e", @"sort" : @-1}, @{@"k" : @"a", @"sort" : @0} ];
+  XCTAssertEqualObjects(FIRQuerySnapshotGetData(snapshot), expected);
+  snapshot = [limitToLastAccumulator awaitEventWithName:@"Snapshot"];
+  expected = @[ @{@"k" : @"a", @"sort" : @0}, @{@"k" : @"e", @"sort" : @-1} ];
+  XCTAssertEqualObjects(FIRQuerySnapshotGetData(snapshot), expected);
+
+  // Unlisten to limitToLast, update a doc, then relisten to limitToLast
+  [limitToLastRegistration remove];
+  [self updateDocumentRef:[collRef documentWithPath:@"a"] data:@{@"k" : @"a", @"sort" : @-2}];
+  [limitToLast addSnapshotListener:[limitToLastAccumulator valueEventHandler]];
+
+  // Verify both queries get expected result.
+  snapshot = [limitAccumulator awaitEventWithName:@"Snapshot"];
+  expected = @[ @{@"k" : @"a", @"sort" : @-2}, @{@"k" : @"e", @"sort" : @-1} ];
+  XCTAssertEqualObjects(FIRQuerySnapshotGetData(snapshot), expected);
+  snapshot = [limitToLastAccumulator awaitEventWithName:@"Snapshot"];
+  expected = @[ @{@"k" : @"e", @"sort" : @-1}, @{@"k" : @"a", @"sort" : @-2} ];
+  XCTAssertEqualObjects(FIRQuerySnapshotGetData(snapshot), expected);
 }
 
 - (void)testKeyOrderIsDescendingForDescendingInequality {