+++ Firestore/core/src/firebase/firestore/remote/remote_event.cc
@@ -26,7 +26,7 @@ namespace firestore {
 namespace remote {
 
 using core::DocumentViewChange;
-using core::Query;
+using core::Target;
 using local::QueryData;
 using local::QueryPurpose;
 using model::DocumentKey;
@@ -215,8 +215,8 @@ void WatchChangeAggregator::HandleExistenceFilter(
 
   absl::optional<QueryData> query_data = QueryDataForActiveTarget(target_id);
   if (query_data) {
-    const Query& query = query_data->query();
-    if (query.IsDocumentQuery()) {
+    const Target& target = query_data->target();
+    if (target.IsDocumentQuery()) {
       if (expected_count == 0) {
         // The existence filter told us the document does not exist. We deduce
         // that this document does not exist and apply a deleted document to our
@@ -224,7 +224,7 @@ void WatchChangeAggregator::HandleExistenceFilter(
         // another query that will raise this document as part of a snapshot
         // until it is resolved, essentially exposing inconsistency between
         // queries.
-        DocumentKey key{query.path()};
+        DocumentKey key{target.path()};
         RemoveDocumentFromTarget(
             target_id, key,
             NoDocument(key, SnapshotVersion::None(),
@@ -256,13 +256,13 @@ RemoteEvent WatchChangeAggregator::CreateRemoteEvent(
 
     absl::optional<QueryData> query_data = QueryDataForActiveTarget(target_id);
     if (query_data) {
-      if (target_state.current() && query_data->query().IsDocumentQuery()) {
+      if (target_state.current() && query_data->target().IsDocumentQuery()) {
         // Document queries for document that don't exist can produce an empty
         // result set. To update our local cache, we synthesize a document
         // delete if we have not previously received the document. This resolves
         // the limbo state of the document, removing it from
         // SyncEngine::limbo_document_refs_.
-        DocumentKey key{query_data->query().path()};
+        DocumentKey key{query_data->target().path()};
         if (pending_document_updates_.find(key) ==
                 pending_document_updates_.end() &&
             !TargetContainsDocument(target_id, key)) {